[Refact] search api refactored
