--- GEON-PPANG-iOS/Global/Network/NewAPI/NewSearchAPI.swift
@@ -1,8 +1,0 @@
-//
-//  SearchAPI.swift
-//  GEON-PPANG-iOS
-//
-//  Created by 이성민 on 1/15/24.
-//
-
-import Foundation
